requirements.txt
requests==2.32.3
pandas==2.2.3
python-dotenv==1.1.0
Flask==3.1.1